+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 3.8.0 (????-??-??)
 
 - Add support for `doctrine/cache` 2.0
+- Bump minimum supported `doctrine/dbal` versions
 
 ## 3.7.2 (2021-04-25)